stock-service/requirements.txt
Flask==3.1.3
APScheduler==3.11.2
requests==2.32.5
psycopg2-binary==2.9.11
python-dotenv==1.2.2
